Restructure files and code based on feedback
