+++ src/AdminOps.ts
@@ -5,7 +5,6 @@ import { baseLogger } from "./logger"
 
 const logger = baseLogger.child({ module: "admin" })
 
-// Now: levelUpdate
 export const setLevel = async ({ uid, level }) => {
   if (Levels.indexOf(level) === -1) {
     const error = `${level} is not a valid user level`
@@ -23,7 +22,6 @@ export const setLevel = async ({ uid, level }) => {
   return user
 }
 
-// Now: mapAddMerchant
 export const addToMap = async ({
   username,
   latitude,
@@ -53,7 +51,6 @@ export const addToMap = async ({
   return !!(await user.save())
 }
 
-// Now: accountUpdateStatus
 export const setAccountStatus = async ({ uid, status }): Promise<typeof User> => {
   const user = await User.findOne({ _id: uid })
