+++ src/entrypoint/graphql-core-server.ts
@@ -1,0 +1,286 @@
+import {
+  stringLength,
+  ValidateDirectiveVisitor,
+  range,
+  pattern,
+} from "@profusion/apollo-validation-directives"
+import dotenv from "dotenv"
+import fs from "fs"
+import { applyMiddleware } from "graphql-middleware"
+import { and, shield } from "graphql-shield"
+import { makeExecutableSchema } from "graphql-tools"
+import moment from "moment"
+import path from "path"
+import { addToMap, setAccountStatus, setLevel } from "../AdminOps"
+import { yamlConfig } from "../config"
+import { getActiveLnd, nodesStats, nodeStats } from "../lndUtils"
+import { getHourlyPrice, getMinBuildNumber } from "../localCache"
+import { sendNotification } from "../notifications/notification"
+import { getCurrentPrice } from "../realtimePrice"
+import { User } from "../schema"
+import { login, requestPhoneCode } from "../text"
+import { Levels, OnboardingEarn, Primitive } from "../types"
+import { WalletFromUsername } from "../walletFactory"
+import { usernameExists } from "../graphql/db"
+
+import { startApolloServer, isAuthenticated, isEditor } from "./graphql-server"
+
+dotenv.config()
+
+const commitHash = process.env.COMMITHASH
+const buildTime = process.env.BUILDTIME
+const helmRevision = process.env.HELMREVISION
+
+const resolvers = {
+  Query: {
+    me: async (_, __, { uid, user }) => {
+      const { phone, username, contacts, language, level } = user
+
+      return {
+        id: uid,
+        level,
+        phone,
+        username,
+        contacts,
+        language,
+      }
+    },
+
+    // legacy, before handling multi currency account
+    wallet: async (_, __, { wallet }) => [
+      {
+        id: "BTC",
+        currency: "BTC",
+        balance: async () => (await wallet.getBalances())["BTC"],
+        transactions: () => wallet.getTransactions(),
+        csv: () => wallet.getStringCsv(),
+      },
+    ],
+
+    // new way to return the balance
+    // balances are distinc between USD and BTC
+    // but transaction are common, because they could have rely both on USD/BTC
+    wallet2: async (_, __, { wallet }) => {
+      const balances = await wallet.getBalances()
+
+      return {
+        transactions: wallet.getTransactions(),
+        balances: wallet.user.currencies.map((item) => ({
+          id: item.id,
+          balance: balances[item.id],
+        })),
+      }
+    },
+    // deprecated
+    nodeStats: async () => {
+      const { lnd } = getActiveLnd()
+      return nodeStats({ lnd })
+    },
+    nodesStats: async () => nodesStats(),
+    buildParameters: async () => {
+      const { minBuildNumber, lastBuildNumber } = await getMinBuildNumber()
+      return {
+        id: lastBuildNumber,
+        commitHash: () => commitHash,
+        buildTime: () => buildTime,
+        helmRevision: () => helmRevision,
+        minBuildNumberAndroid: minBuildNumber,
+        minBuildNumberIos: minBuildNumber,
+        lastBuildNumberAndroid: lastBuildNumber,
+        lastBuildNumberIos: lastBuildNumber,
+      }
+    },
+    prices: async (_, { length = 365 * 24 * 10 }, { logger }) => {
+      const hourly = await getHourlyPrice({ logger })
+
+      // adding the current price as the lat index array
+      // use by the mobile application to convert prices
+      hourly.push({
+        id: moment().unix(),
+        o: getCurrentPrice(),
+      })
+
+      return hourly.splice(-length)
+    },
+    earnList: async (_, __, { user }) => {
+      const response: Record<string, Primitive>[] = []
+      const earned = user?.earn || []
+
+      for (const [id, value] of Object.entries(OnboardingEarn)) {
+        response.push({
+          id,
+          value,
+          completed: earned.findIndex((item) => item === id) !== -1,
+        })
+      }
+
+      return response
+    },
+    getLastOnChainAddress: async (_, __, { wallet }) => ({
+      id: wallet.getLastOnChainAddress(),
+    }),
+    maps: async () => {
+      // TODO: caching
+      const users = await User.find(
+        {
+          title: { $exists: true },
+          coordinate: { $exists: true },
+        },
+        { username: 1, title: 1, coordinate: 1 },
+      )
+
+      return users.map((user) => ({
+        ...user._doc,
+        id: user.username,
+      }))
+    },
+    usernameExists: async (_, { username }) => usernameExists({ username }),
+    getUserDetails: async (_, { uid }) => User.findOne({ _id: uid }),
+    noauthUpdatePendingInvoice: async (_, { hash, username }, { logger }) => {
+      const wallet = await WalletFromUsername({ username, logger })
+      return wallet.updatePendingInvoice({ hash })
+    },
+    getUid: async (_, { username, phone }) => {
+      const { _id: uid } = await User.getUser({ username, phone })
+      return uid
+    },
+    getLevels: () => Levels,
+    getLimits: (_, __, { user }) => {
+      return {
+        oldEnoughForWithdrawal: yamlConfig.limits.oldEnoughForWithdrawal,
+        withdrawal: yamlConfig.limits.withdrawal.level[user.level],
+        onUs: yamlConfig.limits.onUs.level[user.level],
+      }
+    },
+    getWalletFees: () => ({
+      deposit: yamlConfig.fees.deposit,
+    }),
+  },
+  Mutation: {
+    requestPhoneCode: async (_, { phone }, { logger, ip }) => ({
+      success: requestPhoneCode({ phone, logger, ip }),
+    }),
+    login: async (_, { phone, code }, { logger, ip }) => ({
+      token: login({ phone, code, logger, ip }),
+    }),
+    updateUser: async (_, __, { wallet }) => ({
+      setUsername: async ({ username }) => await wallet.setUsername({ username }),
+      setLanguage: async ({ language }) => await wallet.setLanguage({ language }),
+      updateUsername: (input) => wallet.updateUsername(input),
+      updateLanguage: (input) => wallet.updateLanguage(input),
+    }),
+    setLevel: async (_, { uid, level }) => {
+      return setLevel({ uid, level })
+    },
+    updateContact: async (_, __, { user }) => ({
+      setName: async ({ username, name }) => {
+        user.contacts.filter((item) => item.id === username)[0].name = name
+        await user.save()
+        return true
+      },
+    }),
+    noauthAddInvoice: async (_, { username, value }, { logger }) => {
+      const wallet = await WalletFromUsername({ username, logger })
+      return wallet.addInvoice({ selfGenerated: false, value })
+    },
+    invoice: async (_, __, { wallet }) => ({
+      addInvoice: async ({ value, memo }) => wallet.addInvoice({ value, memo }),
+      // FIXME: move to query
+      updatePendingInvoice: async ({ hash }) => wallet.updatePendingInvoice({ hash }),
+      payInvoice: async ({ invoice, amount, memo }) =>
+        wallet.pay({ invoice, amount, memo }),
+      payKeysendUsername: async ({ destination, username, amount, memo }) =>
+        wallet.pay({ destination, username, amount, memo }),
+      getFee: async ({ destination, amount, invoice }) =>
+        wallet.getLightningFee({ destination, amount, invoice }),
+    }),
+    earnCompleted: async (_, { ids }, { wallet }) => wallet.addEarn(ids),
+    onchain: async (_, __, { wallet }) => ({
+      getNewAddress: () => wallet.getOnChainAddress(),
+      pay: ({ address, amount, memo }) => ({
+        success: wallet.onChainPay({ address, amount, memo }),
+      }),
+      payAll: ({ address, memo }) => ({
+        success: wallet.onChainPay({ address, amount: 0, memo, sendAll: true }),
+      }),
+      getFee: ({ address, amount }) => wallet.getOnchainFee({ address, amount }),
+    }),
+    addDeviceToken: async (_, { deviceToken }, { user }) => {
+      user.deviceToken.addToSet(deviceToken)
+      // TODO: check if this is ok to shared a mongoose user instance and mutate it.
+      await user.save()
+      return { success: true }
+    },
+
+    // FIXME test
+    testMessage: async (_, __, { user, logger }) => {
+      // throw new LoggedError("test error")
+      await sendNotification({
+        user,
+        title: "Title",
+        body: `New message sent at ${moment.utc().format("YYYY-MM-DD HH:mm:ss")}`,
+        logger,
+      })
+      return { success: true }
+    },
+    addToMap: async (_, { username, title, latitude, longitude }, { logger }) => {
+      return addToMap({ username, title, latitude, longitude, logger })
+    },
+    setAccountStatus: async (_, { uid, status }) => {
+      return setAccountStatus({ uid, status })
+    },
+  },
+}
+
+export async function startApolloServerForSchema() {
+  const myTypeDefs = fs.readFileSync(path.join(__dirname, "../schema.graphql"), {
+    encoding: "utf8",
+    flag: "r",
+  })
+
+  const execSchema = makeExecutableSchema({
+    typeDefs: [
+      myTypeDefs,
+      ...ValidateDirectiveVisitor.getMissingCommonTypeDefs(),
+      ...range.getTypeDefs(),
+      ...pattern.getTypeDefs(),
+      ...stringLength.getTypeDefs(),
+    ],
+    // @ts-expect-error: TODO
+    schemaDirectives: { pattern, range, stringLength },
+    resolvers,
+  })
+
+  ValidateDirectiveVisitor.addValidationResolversToSchema(execSchema)
+
+  const permissions = shield(
+    {
+      Query: {
+        me: isAuthenticated,
+        wallet: isAuthenticated,
+        wallet2: isAuthenticated,
+        getLastOnChainAddress: isAuthenticated,
+        getUserDetails: and(isAuthenticated, isEditor),
+        getUid: and(isAuthenticated, isEditor),
+        getLevels: and(isAuthenticated, isEditor),
+      },
+      Mutation: {
+        onchain: isAuthenticated,
+        invoice: isAuthenticated,
+        earnCompleted: isAuthenticated,
+        updateUser: isAuthenticated,
+        updateContact: isAuthenticated,
+        addDeviceToken: isAuthenticated,
+        testMessage: isAuthenticated,
+        addToMap: and(isAuthenticated, isEditor),
+        setLevel: and(isAuthenticated, isEditor),
+        setAccountStatus: and(isAuthenticated, isEditor),
+      },
+    },
+    { allowExternalErrors: true },
+  ) // TODO remove to not expose internal error
+
+  const schema = applyMiddleware(execSchema, permissions)
+
+  return startApolloServer({ schema, port: 4000 })
+}
