+++ src/graphql/admin/schema.graphql
